Taxonomic backbone

The taxon names and codes for the modeled taxa in the EVA, with the associated taxon concept information from EuroSL, and parent taxa.

TaxonomicBackbone

Format

A data frame with 6177 rows and 25 columns, the definitions of which are:

taxon_code

The taxon codes used throughout the package, formed from the taxon_name values by coercing all letters to lower and replacing whitespace with underscores.

taxon_name

The taxon names for the modelled taxa, as they appear in the EVA.

taxon_concept

The acccepted taxon concepts for each taxon_name value, following the authority of the EuroSL - see elements::EuroSL

eurosl_status

The taxonomic status of the taxon_name, following the authority of the EuroSL.

eurosl_id

The EuroSL ID code for the taxon_name.

Species

The Species name associated with the taxon_name.

Genus

The parent Genus taxon associated with the taxon_name.

Tribe

The parent Tribe taxon associated with the taxon_name.

Family

The parent Family taxon associated with the taxon_name.

Order

The parent Order taxon associated with the taxon_name.

Superorder

The parent Superorder taxon associated with the taxon_name.

Class

The parent Class taxon associated with the taxon_name.

Subdivision

The parent Subdivision taxon associated with the taxon_name.

Phylum

The parent Phylum taxon associated with the taxon_name.

Root

The parent Root taxon associated with the taxon_name.

Section

The Section name associated with the taxon_name.

Species Aggregate

The Species Aggregate name associated with the taxon_name.

Subclass

The parent Subclass taxon associated with the taxon_name.

Division

The parent Division taxon associated with the taxon_name.

Suprageneric Taxon

The parent Suprageneric Taxon taxon associated with the taxon_name.

Subfamily

The parent Subfamily taxon associated with the taxon_name.

Subspecies

The Subspecies name associated with the taxon_name.

Variety

The Variety name associated with the taxon_name.

Coll. species

The Coll. species name associated with the taxon_name.

n_presences_eva

The number of presences of the taxon (taxon_name) in the EVA, only considering plots with valid coordinates and years.

Details

TaxonomicBackbone