Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa

Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa

Usage

calc_distance_once(
  predictors,
  taxa = ModelledTaxaCodes,
  vars = elements::VariableNames,
  method = "median",
  append = "ids"
)

Arguments

predictors

A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, prec_sm, and optionally taxon

taxa

A vector of strings containing one or more taxa to calculate distances for. See elements::ModelledTaxaCodes.

vars

A vector of variable names, one or more of elements::VariableNames.

method

One of "median" or "mean" indicating whether to measure the euclidean distance from the predictor variables to the median or mean values.

append

A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.

Value

A data frame containing the euclidean distance for each taxon.

Examples

elements::calc_distance_once(predictors = elements::ExampleScenarios[1,])
#>     scenario_code scenario_stage climate_scenario timeslice  scenario_name
#> 1             a.i       baseline       historical      2007 Climate Change
#> 2             a.i       baseline       historical      2007 Climate Change
#> 3             a.i       baseline       historical      2007 Climate Change
#> 4             a.i       baseline       historical      2007 Climate Change
#> 5             a.i       baseline       historical      2007 Climate Change
#> 6             a.i       baseline       historical      2007 Climate Change
#> 7             a.i       baseline       historical      2007 Climate Change
#> 8             a.i       baseline       historical      2007 Climate Change
#> 9             a.i       baseline       historical      2007 Climate Change
#> 10            a.i       baseline       historical      2007 Climate Change
#> 11            a.i       baseline       historical      2007 Climate Change
#> 12            a.i       baseline       historical      2007 Climate Change
#> 13            a.i       baseline       historical      2007 Climate Change
#> 14            a.i       baseline       historical      2007 Climate Change
#> 15            a.i       baseline       historical      2007 Climate Change
#> 16            a.i       baseline       historical      2007 Climate Change
#> 17            a.i       baseline       historical      2007 Climate Change
#> 18            a.i       baseline       historical      2007 Climate Change
#> 19            a.i       baseline       historical      2007 Climate Change
#> 20            a.i       baseline       historical      2007 Climate Change
#> 21            a.i       baseline       historical      2007 Climate Change
#> 22            a.i       baseline       historical      2007 Climate Change
#> 23            a.i       baseline       historical      2007 Climate Change
#> 24            a.i       baseline       historical      2007 Climate Change
#> 25            a.i       baseline       historical      2007 Climate Change
#> 26            a.i       baseline       historical      2007 Climate Change
#> 27            a.i       baseline       historical      2007 Climate Change
#> 28            a.i       baseline       historical      2007 Climate Change
#> 29            a.i       baseline       historical      2007 Climate Change
#> 30            a.i       baseline       historical      2007 Climate Change
#> 31            a.i       baseline       historical      2007 Climate Change
#> 32            a.i       baseline       historical      2007 Climate Change
#> 33            a.i       baseline       historical      2007 Climate Change
#> 34            a.i       baseline       historical      2007 Climate Change
#> 35            a.i       baseline       historical      2007 Climate Change
#> 36            a.i       baseline       historical      2007 Climate Change
#> 37            a.i       baseline       historical      2007 Climate Change
#> 38            a.i       baseline       historical      2007 Climate Change
#> 39            a.i       baseline       historical      2007 Climate Change
#> 40            a.i       baseline       historical      2007 Climate Change
#> 41            a.i       baseline       historical      2007 Climate Change
#> 42            a.i       baseline       historical      2007 Climate Change
#> 43            a.i       baseline       historical      2007 Climate Change
#> 44            a.i       baseline       historical      2007 Climate Change
#> 45            a.i       baseline       historical      2007 Climate Change
#> 46            a.i       baseline       historical      2007 Climate Change
#> 47            a.i       baseline       historical      2007 Climate Change
#> 48            a.i       baseline       historical      2007 Climate Change
#> 49            a.i       baseline       historical      2007 Climate Change
#> 50            a.i       baseline       historical      2007 Climate Change
#> 51            a.i       baseline       historical      2007 Climate Change
#> 52            a.i       baseline       historical      2007 Climate Change
#> 53            a.i       baseline       historical      2007 Climate Change
#> 54            a.i       baseline       historical      2007 Climate Change
#> 55            a.i       baseline       historical      2007 Climate Change
#> 56            a.i       baseline       historical      2007 Climate Change
#> 57            a.i       baseline       historical      2007 Climate Change
#> 58            a.i       baseline       historical      2007 Climate Change
#> 59            a.i       baseline       historical      2007 Climate Change
#> 60            a.i       baseline       historical      2007 Climate Change
#> 61            a.i       baseline       historical      2007 Climate Change
#> 62            a.i       baseline       historical      2007 Climate Change
#> 63            a.i       baseline       historical      2007 Climate Change
#> 64            a.i       baseline       historical      2007 Climate Change
#> 65            a.i       baseline       historical      2007 Climate Change
#> 66            a.i       baseline       historical      2007 Climate Change
#> 67            a.i       baseline       historical      2007 Climate Change
#> 68            a.i       baseline       historical      2007 Climate Change
#> 69            a.i       baseline       historical      2007 Climate Change
#> 70            a.i       baseline       historical      2007 Climate Change
#> 71            a.i       baseline       historical      2007 Climate Change
#> 72            a.i       baseline       historical      2007 Climate Change
#> 73            a.i       baseline       historical      2007 Climate Change
#> 74            a.i       baseline       historical      2007 Climate Change
#> 75            a.i       baseline       historical      2007 Climate Change
#> 76            a.i       baseline       historical      2007 Climate Change
#> 77            a.i       baseline       historical      2007 Climate Change
#> 78            a.i       baseline       historical      2007 Climate Change
#> 79            a.i       baseline       historical      2007 Climate Change
#> 80            a.i       baseline       historical      2007 Climate Change
#> 81            a.i       baseline       historical      2007 Climate Change
#> 82            a.i       baseline       historical      2007 Climate Change
#> 83            a.i       baseline       historical      2007 Climate Change
#> 84            a.i       baseline       historical      2007 Climate Change
#> 85            a.i       baseline       historical      2007 Climate Change
#> 86            a.i       baseline       historical      2007 Climate Change
#> 87            a.i       baseline       historical      2007 Climate Change
#> 88            a.i       baseline       historical      2007 Climate Change
#> 89            a.i       baseline       historical      2007 Climate Change
#> 90            a.i       baseline       historical      2007 Climate Change
#> 91            a.i       baseline       historical      2007 Climate Change
#> 92            a.i       baseline       historical      2007 Climate Change
#> 93            a.i       baseline       historical      2007 Climate Change
#> 94            a.i       baseline       historical      2007 Climate Change
#> 95            a.i       baseline       historical      2007 Climate Change
#> 96            a.i       baseline       historical      2007 Climate Change
#> 97            a.i       baseline       historical      2007 Climate Change
#> 98            a.i       baseline       historical      2007 Climate Change
#> 99            a.i       baseline       historical      2007 Climate Change
#> 100           a.i       baseline       historical      2007 Climate Change
#> 101           a.i       baseline       historical      2007 Climate Change
#> 102           a.i       baseline       historical      2007 Climate Change
#> 103           a.i       baseline       historical      2007 Climate Change
#> 104           a.i       baseline       historical      2007 Climate Change
#> 105           a.i       baseline       historical      2007 Climate Change
#> 106           a.i       baseline       historical      2007 Climate Change
#> 107           a.i       baseline       historical      2007 Climate Change
#> 108           a.i       baseline       historical      2007 Climate Change
#> 109           a.i       baseline       historical      2007 Climate Change
#> 110           a.i       baseline       historical      2007 Climate Change
#> 111           a.i       baseline       historical      2007 Climate Change
#> 112           a.i       baseline       historical      2007 Climate Change
#> 113           a.i       baseline       historical      2007 Climate Change
#> 114           a.i       baseline       historical      2007 Climate Change
#> 115           a.i       baseline       historical      2007 Climate Change
#> 116           a.i       baseline       historical      2007 Climate Change
#> 117           a.i       baseline       historical      2007 Climate Change
#> 118           a.i       baseline       historical      2007 Climate Change
#> 119           a.i       baseline       historical      2007 Climate Change
#> 120           a.i       baseline       historical      2007 Climate Change
#> 121           a.i       baseline       historical      2007 Climate Change
#> 122           a.i       baseline       historical      2007 Climate Change
#> 123           a.i       baseline       historical      2007 Climate Change
#> 124           a.i       baseline       historical      2007 Climate Change
#> 125           a.i       baseline       historical      2007 Climate Change
#> 126           a.i       baseline       historical      2007 Climate Change
#> 127           a.i       baseline       historical      2007 Climate Change
#> 128           a.i       baseline       historical      2007 Climate Change
#> 129           a.i       baseline       historical      2007 Climate Change
#> 130           a.i       baseline       historical      2007 Climate Change
#> 131           a.i       baseline       historical      2007 Climate Change
#> 132           a.i       baseline       historical      2007 Climate Change
#> 133           a.i       baseline       historical      2007 Climate Change
#> 134           a.i       baseline       historical      2007 Climate Change
#> 135           a.i       baseline       historical      2007 Climate Change
#> 136           a.i       baseline       historical      2007 Climate Change
#> 137           a.i       baseline       historical      2007 Climate Change
#> 138           a.i       baseline       historical      2007 Climate Change
#> 139           a.i       baseline       historical      2007 Climate Change
#> 140           a.i       baseline       historical      2007 Climate Change
#> 141           a.i       baseline       historical      2007 Climate Change
#> 142           a.i       baseline       historical      2007 Climate Change
#>                                      taxon_code distance
#> 1                                    abies_alba 3.290727
#> 2                           abies_borisii-regis 3.203137
#> 3                             abies_cephalonica 3.383786
#> 4                abies_cilicica_subsp._isaurica 3.328743
#> 5                                 abies_grandis 2.559653
#> 6                            abies_nordmanniana 2.444428
#> 7        abies_nordmanniana_subsp._equi-trojani 2.789854
#> 8        abies_nordmanniana_subsp._nordmanniana 3.352738
#> 9                                 abies_procera 2.446669
#> 10                         abietinella_abietina 2.608063
#> 11                         abutilon_theophrasti 2.991759
#> 12                              acacia_dealbata 2.835607
#> 13                               acacia_saligna 3.033461
#> 14                           acalypha_virginica 3.038219
#> 15                        acantholimon_acerosum 2.877894
#> 16  acantholimon_acerosum_subsp._brachystachyum 2.815904
#> 17                         acantholimon_echinus 2.942183
#> 18                       acantholimon_puberulum 2.751694
#> 19        acantholimon_puberulum_var._puberulum 2.988561
#> 20                        acantholimon_ulicinum 3.360145
#> 21        acantholimon_ulicinum_subsp._ulicinum 3.255066
#> 22                        acantholimon_venestum 2.901526
#> 23                        acantholimon_venustum 2.934785
#> 24          acantholimon_venustum_var._venustum 3.499293
#> 25                            acanthus_hirsutus 2.825348
#> 26                          acanthus_hungaricus 2.455483
#> 27                              acanthus_mollis 3.183755
#> 28                acanthus_mollis_subsp._mollis 2.631565
#> 29                            acanthus_spinosus 2.621390
#> 30                           acarospora_fuscata 2.555449
#> 31                               acer_campestre 2.959782
#> 32              acer_campestre_subsp._campestre 2.707919
#> 33                            acer_cappadocicum 3.426917
#> 34                             acer_heldreichii 3.362432
#> 35                                acer_hyrcanum 2.894737
#> 36                                 acer_lobelii 3.600290
#> 37                          acer_monspessulanum 2.663544
#> 38      acer_monspessulanum_subsp._microphyllum 3.386163
#> 39    acer_monspessulanum_subsp._monspessulanum 2.987290
#> 40                                 acer_negundo 2.945201
#> 41                               acer_obtusatum 3.170468
#> 42                                  acer_opalus 2.918517
#> 43                             acer_platanoides 3.234520
#> 44                          acer_pseudoplatanus 3.176335
#> 45                            acer_sempervirens 3.409572
#> 46                               acer_tataricum 2.702094
#> 47                            acer_trautvetteri 3.405717
#> 48                        achillea_abrotanoides 2.626071
#> 49                        achillea_ageratifolia 3.045371
#> 50                            achillea_ageratum 3.089178
#> 51            achillea_aleppica_subsp._aleppica 3.044897
#> 52         achillea_aleppica_subsp._zederbaueri 3.265013
#> 53                             achillea_arabica 2.796634
#> 54                       achillea_aspleniifolia 2.324600
#> 55                              achillea_atrata 2.401211
#> 56       achillea_barrelieri_subsp._mucronulata 3.841817
#> 57                           achillea_clavennae 2.544495
#> 58                            achillea_clusiana 2.362763
#> 59                          achillea_clypeolata 2.713711
#> 60                           achillea_coarctata 2.950071
#> 61                             achillea_collina 2.525134
#> 62                        achillea_crithmifolia 2.600154
#> 63                             achillea_distans 2.381165
#> 64              achillea_distans_subsp._stricta 2.505953
#> 65        achillea_distans_subsp._tanacetifolia 2.392164
#> 66          achillea_erba-rotta_subsp._moschata 2.647539
#> 67                              achillea_euxina 2.706652
#> 68                             achillea_falcata 2.674280
#> 69                             achillea_fraasii 3.412639
#> 70                         achillea_grandifolia 2.873950
#> 71                         achillea_holosericea 3.327259
#> 72                            achillea_inundata 2.764236
#> 73            achillea_kotschyi_subsp._kotschyi 3.630286
#> 74                         achillea_leptophylla 2.983601
#> 75                           achillea_ligustica 3.120215
#> 76                           achillea_lingulata 2.464730
#> 77                         achillea_macrophylla 3.499823
#> 78                            achillea_maritima 3.017395
#> 79            achillea_maritima_subsp._maritima 3.274271
#> 80                           achillea_micrantha 2.948348
#> 81                         achillea_millefolium 2.424338
#> 82                    achillea_millefolium_aggr 2.507676
#> 83      achillea_millefolium_subsp._millefolium 2.461473
#> 84         achillea_millefolium_subsp._sudetica 2.491517
#> 85                                achillea_nana 2.843504
#> 86                             achillea_nobilis 2.650581
#> 87          achillea_nobilis_subsp._neilreichii 2.691049
#> 88              achillea_nobilis_subsp._nobilis 2.883282
#> 89                          achillea_ochroleuca 3.021631
#> 90                             achillea_odorata 3.311568
#> 91                             achillea_oxyloba 2.556651
#> 92              achillea_oxyloba_subsp._schurii 2.215693
#> 93                           achillea_pannonica 2.618875
#> 94                             achillea_phrygia 2.789143
#> 95                           achillea_pratensis 2.566658
#> 96                            achillea_ptarmica 2.312536
#> 97                           achillea_pyrenaica 2.568272
#> 98                           achillea_roseoalba 2.808413
#> 99                         achillea_salicifolia 2.609699
#> 100    achillea_santolinoides_subsp._wilhelmsii 3.186791
#> 101                            achillea_setacea 2.673112
#> 102                         achillea_sintenisii 3.293599
#> 103                           achillea_styriaca 2.973419
#> 104                        achillea_teretifolia 2.815730
#> 105                          achillea_tomentosa 2.970291
#> 106                          achillea_virescens 2.545221
#> 107           achillea_virescens_subsp._tenorei 3.738573
#> 108                       achnatherum_bromoides 3.056308
#> 109                   achnatherum_calamagrostis 2.541247
#> 110                      aconitum_angustifolium 2.570210
#> 111                            aconitum_anthora 2.566946
#> 112                            aconitum_degenii 3.212914
#> 113                             aconitum_firmum 2.536360
#> 114                         aconitum_lycoctonum 3.172342
#> 115       aconitum_lycoctonum_subsp._lycoctonum 3.159415
#> 116     aconitum_lycoctonum_subsp._neapolitanum 2.598042
#> 117         aconitum_lycoctonum_subsp._vulparia 3.476391
#> 118                         aconitum_moldavicum 3.332661
#> 119                           aconitum_napellus 2.636463
#> 120                      aconitum_napellus_aggr 3.044166
#> 121            aconitum_napellus_subsp._vulgare 3.149239
#> 122                           aconitum_plicatum 2.681721
#> 123                           aconitum_tauricum 2.347740
#> 124                            aconitum_toxicum 2.609242
#> 125                         aconitum_variegatum 2.936283
#> 126                    aconitum_variegatum_aggr 2.996366
#> 127          aconitum_variegatum_subsp._nasutum 3.088891
#> 128                          aconogonon_alpinum 2.468692
#> 129                              acorus_calamus 2.715531
#> 130                  acrocarpe_mossen_(overige) 2.445630
#> 131                             actaea_europaea 3.080377
#> 132                              actaea_spicata 3.446216
#> 133                     adenocarpus_complicatus 2.411959
#> 134                       adenophora_liliifolia 2.420457
#> 135                       adenostyles_alliariae 3.052103
#> 136      adenostyles_alliariae_subsp._alliariae 3.466476
#> 137                          adenostyles_alpina 3.018982
#> 138            adenostyles_alpina_subsp._alpina 3.202158
#> 139            adenostyles_glabra_subsp._glabra 3.906979
#> 140                     adenostyles_leucophylla 2.699815
#> 141                   adiantum_capillus-veneris 2.769876
#> 142                           adonis_aestivalis 2.604191
#>  [ reached 'max' / getOption("max.print") -- omitted 9347 rows ]