Calculate the euclidean distance between a set of environmental predictor variables and mutliple taxa

Calculate the euclidean distance between a set of environmental predictor variables and mutliple taxa

Usage

calc_distance(
  predictors,
  taxa = elements::TaxonomicBackbone$taxon_code,
  vars = elements::VariableNames,
  method = "median",
  append = "ids"
)

Arguments

predictors

A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17, and optionally taxon

taxa

A vector of strings containing one or more taxa to calculate distances for. See elements::TaxonomicBackbone.

vars

A vector of variable names, one or more of elements::VariableNames.

method

One of "median" or "mean" indicating whether to measure the euclidean distance from the predictor variables to the median or mean values.

append

A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.

Value

A data frame containing the euclidean distance for each taxon.

Examples

elements::calc_distance(predictors = elements::ExampleScenarios)
#>       scenario timeslice scenario_code
#> 1.1   Baseline      2007             a
#> 1.2   Baseline      2007             a
#> 1.3   Baseline      2007             a
#> 1.4   Baseline      2007             a
#> 1.5   Baseline      2007             a
#> 1.6   Baseline      2007             a
#> 1.7   Baseline      2007             a
#> 1.8   Baseline      2007             a
#> 1.9   Baseline      2007             a
#> 1.10  Baseline      2007             a
#> 1.11  Baseline      2007             a
#> 1.12  Baseline      2007             a
#> 1.13  Baseline      2007             a
#> 1.14  Baseline      2007             a
#> 1.15  Baseline      2007             a
#> 1.16  Baseline      2007             a
#> 1.17  Baseline      2007             a
#> 1.18  Baseline      2007             a
#> 1.19  Baseline      2007             a
#> 1.20  Baseline      2007             a
#> 1.21  Baseline      2007             a
#> 1.22  Baseline      2007             a
#> 1.23  Baseline      2007             a
#> 1.24  Baseline      2007             a
#> 1.25  Baseline      2007             a
#> 1.26  Baseline      2007             a
#> 1.27  Baseline      2007             a
#> 1.28  Baseline      2007             a
#> 1.29  Baseline      2007             a
#> 1.30  Baseline      2007             a
#> 1.31  Baseline      2007             a
#> 1.32  Baseline      2007             a
#> 1.33  Baseline      2007             a
#> 1.34  Baseline      2007             a
#> 1.35  Baseline      2007             a
#> 1.36  Baseline      2007             a
#> 1.37  Baseline      2007             a
#> 1.38  Baseline      2007             a
#> 1.39  Baseline      2007             a
#> 1.40  Baseline      2007             a
#> 1.41  Baseline      2007             a
#> 1.42  Baseline      2007             a
#> 1.43  Baseline      2007             a
#> 1.44  Baseline      2007             a
#> 1.45  Baseline      2007             a
#> 1.46  Baseline      2007             a
#> 1.47  Baseline      2007             a
#> 1.48  Baseline      2007             a
#> 1.49  Baseline      2007             a
#> 1.50  Baseline      2007             a
#> 1.51  Baseline      2007             a
#> 1.52  Baseline      2007             a
#> 1.53  Baseline      2007             a
#> 1.54  Baseline      2007             a
#> 1.55  Baseline      2007             a
#> 1.56  Baseline      2007             a
#> 1.57  Baseline      2007             a
#> 1.58  Baseline      2007             a
#> 1.59  Baseline      2007             a
#> 1.60  Baseline      2007             a
#> 1.61  Baseline      2007             a
#> 1.62  Baseline      2007             a
#> 1.63  Baseline      2007             a
#> 1.64  Baseline      2007             a
#> 1.65  Baseline      2007             a
#> 1.66  Baseline      2007             a
#> 1.67  Baseline      2007             a
#> 1.68  Baseline      2007             a
#> 1.69  Baseline      2007             a
#> 1.70  Baseline      2007             a
#> 1.71  Baseline      2007             a
#> 1.72  Baseline      2007             a
#> 1.73  Baseline      2007             a
#> 1.74  Baseline      2007             a
#> 1.75  Baseline      2007             a
#> 1.76  Baseline      2007             a
#> 1.77  Baseline      2007             a
#> 1.78  Baseline      2007             a
#> 1.79  Baseline      2007             a
#> 1.80  Baseline      2007             a
#> 1.81  Baseline      2007             a
#> 1.82  Baseline      2007             a
#> 1.83  Baseline      2007             a
#> 1.84  Baseline      2007             a
#> 1.85  Baseline      2007             a
#> 1.86  Baseline      2007             a
#> 1.87  Baseline      2007             a
#> 1.88  Baseline      2007             a
#> 1.89  Baseline      2007             a
#> 1.90  Baseline      2007             a
#> 1.91  Baseline      2007             a
#> 1.92  Baseline      2007             a
#> 1.93  Baseline      2007             a
#> 1.94  Baseline      2007             a
#> 1.95  Baseline      2007             a
#> 1.96  Baseline      2007             a
#> 1.97  Baseline      2007             a
#> 1.98  Baseline      2007             a
#> 1.99  Baseline      2007             a
#> 1.100 Baseline      2007             a
#> 1.101 Baseline      2007             a
#> 1.102 Baseline      2007             a
#> 1.103 Baseline      2007             a
#> 1.104 Baseline      2007             a
#> 1.105 Baseline      2007             a
#> 1.106 Baseline      2007             a
#> 1.107 Baseline      2007             a
#> 1.108 Baseline      2007             a
#> 1.109 Baseline      2007             a
#> 1.110 Baseline      2007             a
#> 1.111 Baseline      2007             a
#> 1.112 Baseline      2007             a
#> 1.113 Baseline      2007             a
#> 1.114 Baseline      2007             a
#> 1.115 Baseline      2007             a
#> 1.116 Baseline      2007             a
#> 1.117 Baseline      2007             a
#> 1.118 Baseline      2007             a
#> 1.119 Baseline      2007             a
#> 1.120 Baseline      2007             a
#> 1.121 Baseline      2007             a
#> 1.122 Baseline      2007             a
#> 1.123 Baseline      2007             a
#> 1.124 Baseline      2007             a
#> 1.125 Baseline      2007             a
#> 1.126 Baseline      2007             a
#> 1.127 Baseline      2007             a
#> 1.128 Baseline      2007             a
#> 1.129 Baseline      2007             a
#> 1.130 Baseline      2007             a
#> 1.131 Baseline      2007             a
#> 1.132 Baseline      2007             a
#> 1.133 Baseline      2007             a
#> 1.134 Baseline      2007             a
#> 1.135 Baseline      2007             a
#> 1.136 Baseline      2007             a
#> 1.137 Baseline      2007             a
#> 1.138 Baseline      2007             a
#> 1.139 Baseline      2007             a
#> 1.140 Baseline      2007             a
#> 1.141 Baseline      2007             a
#> 1.142 Baseline      2007             a
#> 1.143 Baseline      2007             a
#> 1.144 Baseline      2007             a
#> 1.145 Baseline      2007             a
#> 1.146 Baseline      2007             a
#> 1.147 Baseline      2007             a
#> 1.148 Baseline      2007             a
#> 1.149 Baseline      2007             a
#> 1.150 Baseline      2007             a
#> 1.151 Baseline      2007             a
#> 1.152 Baseline      2007             a
#> 1.153 Baseline      2007             a
#> 1.154 Baseline      2007             a
#> 1.155 Baseline      2007             a
#> 1.156 Baseline      2007             a
#> 1.157 Baseline      2007             a
#> 1.158 Baseline      2007             a
#> 1.159 Baseline      2007             a
#> 1.160 Baseline      2007             a
#> 1.161 Baseline      2007             a
#> 1.162 Baseline      2007             a
#> 1.163 Baseline      2007             a
#> 1.164 Baseline      2007             a
#> 1.165 Baseline      2007             a
#> 1.166 Baseline      2007             a
#> 1.167 Baseline      2007             a
#> 1.168 Baseline      2007             a
#> 1.169 Baseline      2007             a
#> 1.170 Baseline      2007             a
#> 1.171 Baseline      2007             a
#> 1.172 Baseline      2007             a
#> 1.173 Baseline      2007             a
#> 1.174 Baseline      2007             a
#> 1.175 Baseline      2007             a
#> 1.176 Baseline      2007             a
#> 1.177 Baseline      2007             a
#> 1.178 Baseline      2007             a
#> 1.179 Baseline      2007             a
#> 1.180 Baseline      2007             a
#> 1.181 Baseline      2007             a
#> 1.182 Baseline      2007             a
#> 1.183 Baseline      2007             a
#> 1.184 Baseline      2007             a
#> 1.185 Baseline      2007             a
#> 1.186 Baseline      2007             a
#> 1.187 Baseline      2007             a
#> 1.188 Baseline      2007             a
#> 1.189 Baseline      2007             a
#> 1.190 Baseline      2007             a
#> 1.191 Baseline      2007             a
#> 1.192 Baseline      2007             a
#> 1.193 Baseline      2007             a
#> 1.194 Baseline      2007             a
#> 1.195 Baseline      2007             a
#> 1.196 Baseline      2007             a
#> 1.197 Baseline      2007             a
#> 1.198 Baseline      2007             a
#> 1.199 Baseline      2007             a
#> 1.200 Baseline      2007             a
#>                                      taxon_code distance
#> 1.1                                  abies_alba 3.636952
#> 1.2                         abies_borisii-regis 3.406591
#> 1.3                           abies_cephalonica 3.123997
#> 1.4              abies_cilicica_subsp._isaurica 2.899396
#> 1.5                               abies_grandis 3.219113
#> 1.6      abies_nordmanniana_subsp._equi-trojani 3.387160
#> 1.7      abies_nordmanniana_subsp._nordmanniana 3.683923
#> 1.8                        abietinella_abietina 2.860623
#> 1.9                        abutilon_theophrasti 2.913475
#> 1.10                      acantholimon_acerosum 2.833773
#> 1.11      acantholimon_puberulum_var._puberulum 2.809365
#> 1.12                      acantholimon_ulicinum 2.937199
#> 1.13      acantholimon_ulicinum_subsp._ulicinum 2.802648
#> 1.14        acantholimon_venustum_var._venustum 2.850136
#> 1.15                        acanthus_hungaricus 3.068189
#> 1.16                            acanthus_mollis 3.251038
#> 1.17              acanthus_mollis_subsp._mollis 3.257391
#> 1.18                          acanthus_spinosus 3.277638
#> 1.19                             acer_campestre 3.316213
#> 1.20            acer_campestre_subsp._campestre 3.126469
#> 1.21                            acer_granatense 3.012411
#> 1.22                           acer_heldreichii 3.591992
#> 1.23                              acer_hyrcanum 3.231464
#> 1.24                        acer_monspessulanum 3.140927
#> 1.25  acer_monspessulanum_subsp._monspessulanum 3.287096
#> 1.26                               acer_negundo 3.007178
#> 1.27                             acer_obtusatum 3.444014
#> 1.28                                acer_opalus 3.297285
#> 1.29                           acer_platanoides 3.406773
#> 1.30                        acer_pseudoplatanus 3.413035
#> 1.31                          acer_sempervirens 3.212860
#> 1.32                             acer_tataricum 3.182642
#> 1.33                          acer_trautvetteri 3.599463
#> 1.34                           achillea_arabica 2.830811
#> 1.35                     achillea_aspleniifolia 2.830588
#> 1.36                            achillea_atrata 3.653152
#> 1.37                         achillea_clavennae 3.598935
#> 1.38                          achillea_clusiana 3.314449
#> 1.39                        achillea_clypeolata 2.927989
#> 1.40                         achillea_coarctata 2.901220
#> 1.41                           achillea_collina 2.868363
#> 1.42                      achillea_crithmifolia 2.878634
#> 1.43                           achillea_distans 3.238368
#> 1.44            achillea_distans_subsp._stricta 3.191846
#> 1.45        achillea_erba-rotta_subsp._moschata 3.681318
#> 1.46                       achillea_grandifolia 3.079541
#> 1.47                       achillea_holosericea 3.103055
#> 1.48                         achillea_ligustica 2.942638
#> 1.49                       achillea_macrophylla 3.971991
#> 1.50                          achillea_maritima 3.234618
#> 1.51          achillea_maritima_subsp._maritima 3.259122
#> 1.52                         achillea_micrantha 2.801288
#> 1.53                     achillea_micranthoides 2.819302
#> 1.54                       achillea_millefolium 2.842956
#> 1.55                 achillea_millefolium_aggr. 2.889183
#> 1.56    achillea_millefolium_subsp._millefolium 2.797223
#> 1.57       achillea_millefolium_subsp._sudetica 3.297393
#> 1.58                              achillea_nana 3.586640
#> 1.59                           achillea_nobilis 2.849483
#> 1.60        achillea_nobilis_subsp._neilreichii 2.851591
#> 1.61            achillea_nobilis_subsp._nobilis 2.854779
#> 1.62                           achillea_odorata 2.916762
#> 1.63                           achillea_oxyloba 3.569481
#> 1.64            achillea_oxyloba_subsp._schurii 3.336302
#> 1.65                         achillea_pannonica 2.851478
#> 1.66                          achillea_ptarmica 2.858622
#> 1.67                         achillea_pyrenaica 3.039153
#> 1.68                       achillea_salicifolia 2.838002
#> 1.69   achillea_santolinoides_subsp._wilhelmsii 2.832791
#> 1.70                           achillea_setacea 2.868584
#> 1.71                         achillea_tomentosa 2.985516
#> 1.72                         achillea_virescens 3.094862
#> 1.73                      achnatherum_bromoides 3.189521
#> 1.74                  achnatherum_calamagrostis 3.067461
#> 1.75                     aconitum_angustifolium 3.640520
#> 1.76                           aconitum_anthora 3.046668
#> 1.77                           aconitum_degenii 3.904280
#> 1.78                            aconitum_firmum 3.365457
#> 1.79                        aconitum_lycoctonum 3.800489
#> 1.80      aconitum_lycoctonum_subsp._lycoctonum 3.588196
#> 1.81    aconitum_lycoctonum_subsp._neapolitanum 3.562313
#> 1.82        aconitum_lycoctonum_subsp._vulparia 3.808958
#> 1.83                        aconitum_moldavicum 3.520008
#> 1.84                          aconitum_napellus 3.365612
#> 1.85                    aconitum_napellus_aggr. 3.831748
#> 1.86           aconitum_napellus_subsp._vulgare 3.509561
#> 1.87                          aconitum_plicatum 3.312191
#> 1.88                          aconitum_tauricum 3.522273
#> 1.89                        aconitum_variegatum 3.374926
#> 1.90                         aconogonon_alpinum 3.168574
#> 1.91                             acorus_calamus 2.883851
#> 1.92                            actaea_europaea 3.311772
#> 1.93                             actaea_spicata 3.610888
#> 1.94                    adenocarpus_complicatus 2.997037
#> 1.95                     adenocarpus_telonensis 3.065979
#> 1.96                      adenostyles_alliariae 3.744186
#> 1.97     adenostyles_alliariae_subsp._alliariae 3.886905
#> 1.98                         adenostyles_alpina 3.881803
#> 1.99           adenostyles_alpina_subsp._alpina 3.547925
#> 1.100       adenostyles_alpina_subsp._pyrenaica 3.555039
#> 1.101          adenostyles_glabra_subsp._glabra 3.797002
#> 1.102                   adenostyles_leucophylla 3.338009
#> 1.103                 adiantum_capillus-veneris 3.240927
#> 1.104                         adonis_aestivalis 2.865460
#> 1.105                              adonis_annua 3.046768
#> 1.106                            adonis_flammea 2.869221
#> 1.107                         adonis_microcarpa 3.013274
#> 1.108                           adonis_vernalis 2.902688
#> 1.109                       adoxa_moschatellina 3.442337
#> 1.110                       aegilops_biuncialis 3.122875
#> 1.111                          aegilops_caudata 3.250102
#> 1.112                       aegilops_cylindrica 2.892887
#> 1.113                       aegilops_geniculata 3.021595
#> 1.114                         aegilops_neglecta 2.933335
#> 1.115                      aegilops_triuncialis 2.947133
#> 1.116   aegilops_triuncialis_subsp._triuncialis 2.963894
#> 1.117                      aegilops_umbellulata 2.932909
#> 1.118               aegonychon_purpurocaeruleum 3.178334
#> 1.119                     aegopodium_podagraria 3.225339
#> 1.120                      aeluropus_littoralis 2.967920
#> 1.121                    aesculus_hippocastanum 3.280604
#> 1.122                       aethionema_arabicum 2.843364
#> 1.123                       aethionema_saxatile 3.028566
#> 1.124        aethionema_saxatile_subsp._graecum 3.058361
#> 1.125                          aethusa_cynapium 2.883293
#> 1.126          aethusa_cynapium_subsp._cynapium 2.826450
#> 1.127                       agrimonia_eupatoria 2.887922
#> 1.128      agrimonia_eupatoria_subsp._eupatoria 2.825901
#> 1.129                         agrimonia_procera 2.949042
#> 1.130                       agropyron_cristatum 2.888303
#> 1.131       agropyron_cristatum_subsp._brandzae 2.904341
#> 1.132     agropyron_cristatum_subsp._pectinatum 2.865567
#> 1.133                      agropyron_dasyanthum 2.788287
#> 1.134                        agrostemma_githago 2.844972
#> 1.135                           agrostis_alpina 3.529333
#> 1.136                           agrostis_canina 2.829460
#> 1.137                     agrostis_canina_aggr. 2.847522
#> 1.138                      agrostis_canina_s.l. 3.058093
#> 1.139                       agrostis_capillaris 2.836853
#> 1.140                       agrostis_castellana 2.885383
#> 1.141                         agrostis_curtisii 2.855063
#> 1.142                         agrostis_gigantea 2.828203
#> 1.143         agrostis_gigantea_subsp._maeotica 2.774186
#> 1.144                        agrostis_hesperica 2.771401
#> 1.145                        agrostis_mertensii 3.144778
#> 1.146                        agrostis_rupestris 3.313447
#> 1.147                      agrostis_schleicheri 3.046642
#> 1.148                     agrostis_schraderiana 3.467704
#> 1.149                      agrostis_stolonifera 2.846342
#> 1.150                agrostis_stolonifera_aggr. 2.883735
#> 1.151      agrostis_stolonifera_subsp._maritima 3.002495
#> 1.152                         agrostis_vinealis 2.810356
#> 1.153                   agrostis_x_fouilladeana 2.924155
#> 1.154                       ailanthus_altissima 2.985830
#> 1.155                         aira_caryophyllea 2.837323
#> 1.156     aira_caryophyllea_subsp._caryophyllea 2.825524
#> 1.157                            aira_cupaniana 3.103830
#> 1.158                        aira_elegantissima 3.006436
#> 1.159                              aira_praecox 2.800826
#> 1.160                              aira_tenorei 3.078650
#> 1.161                          aira_uniaristata 3.064657
#> 1.162                         ajuga_chamaepitys 2.922216
#> 1.163             ajuga_chamaepitys_subsp._chia 2.871643
#> 1.164                          ajuga_genevensis 2.890729
#> 1.165                                 ajuga_iva 3.114191
#> 1.166                           ajuga_laxmannii 2.919480
#> 1.167                          ajuga_orientalis 2.885406
#> 1.168                         ajuga_pyramidalis 3.193739
#> 1.169                             ajuga_reptans 3.203512
#> 1.170                             alcea_biennis 3.695590
#> 1.171                      alchemilla_acutiloba 2.965728
#> 1.172                       alchemilla_alpigena 3.122777
#> 1.173                         alchemilla_alpina 3.346503
#> 1.174                       alchemilla_anisiaca 3.621825
#> 1.175                    alchemilla_catalaunica 3.250545
#> 1.176                      alchemilla_caucasica 3.351831
#> 1.177                       alchemilla_colorata 3.183223
#> 1.178                      alchemilla_conjuncta 3.739237
#> 1.179                alchemilla_conjuncta_aggr. 3.629804
#> 1.180                       alchemilla_coriacea 3.287342
#> 1.181                 alchemilla_coriacea_aggr. 3.451742
#> 1.182                        alchemilla_crinita 3.101606
#> 1.183                     alchemilla_filicaulis 3.077561
#> 1.184                          alchemilla_fissa 3.398253
#> 1.185                     alchemilla_flabellata 3.170051
#> 1.186                         alchemilla_glabra 3.114937
#> 1.187                   alchemilla_glabra_aggr. 3.511190
#> 1.188                    alchemilla_glaucescens 2.895560
#> 1.189                       alchemilla_hoppeana 3.364724
#> 1.190                        alchemilla_hybrida 3.126706
#> 1.191                  alchemilla_hybrida_aggr. 3.412525
#> 1.192                         alchemilla_micans 2.966234
#> 1.193                      alchemilla_monticola 2.971383
#> 1.194                   alchemilla_pentaphyllea 3.579574
#> 1.195       alchemilla_plicatula_var._plicatula 3.366188
#> 1.196                      alchemilla_saxatilis 3.119380
#> 1.197                     alchemilla_subcrenata 2.999995
#> 1.198                       alchemilla_vulgaris 3.050119
#> 1.199                 alchemilla_vulgaris_aggr. 3.031662
#> 1.200                   alchemilla_xanthochlora 3.029799
#>  [ reached 'max' / getOption("max.print") -- omitted 172206 rows ]