Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa

Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa

Usage

calc_distance_once(
  predictors,
  taxa = elements::TaxonomicBackbone$taxon_code,
  vars = elements::VariableNames,
  method = "median",
  append = "ids"
)

Arguments

predictors

A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, bio17, and optionally taxon

taxa

A vector of strings containing one or more taxa to calculate distances for. See elements::TaxonomicBackbone.

vars

A vector of variable names, one or more of elements::VariableNames.

method

One of "median" or "mean" indicating whether to measure the euclidean distance from the predictor variables to the median or mean values.

append

A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.

Value

A data frame containing the euclidean distance for each taxon.

Examples

elements::calc_distance_once(predictors = elements::ExampleScenarios[1,])
#>     scenario timeslice scenario_code                                taxon_code
#> 1   Baseline      2007             a                                abies_alba
#> 2   Baseline      2007             a                       abies_borisii-regis
#> 3   Baseline      2007             a                         abies_cephalonica
#> 4   Baseline      2007             a            abies_cilicica_subsp._isaurica
#> 5   Baseline      2007             a                             abies_grandis
#> 6   Baseline      2007             a    abies_nordmanniana_subsp._equi-trojani
#> 7   Baseline      2007             a    abies_nordmanniana_subsp._nordmanniana
#> 8   Baseline      2007             a                      abietinella_abietina
#> 9   Baseline      2007             a                      abutilon_theophrasti
#> 10  Baseline      2007             a                     acantholimon_acerosum
#> 11  Baseline      2007             a     acantholimon_puberulum_var._puberulum
#> 12  Baseline      2007             a                     acantholimon_ulicinum
#> 13  Baseline      2007             a     acantholimon_ulicinum_subsp._ulicinum
#> 14  Baseline      2007             a       acantholimon_venustum_var._venustum
#> 15  Baseline      2007             a                       acanthus_hungaricus
#> 16  Baseline      2007             a                           acanthus_mollis
#> 17  Baseline      2007             a             acanthus_mollis_subsp._mollis
#> 18  Baseline      2007             a                         acanthus_spinosus
#> 19  Baseline      2007             a                            acer_campestre
#> 20  Baseline      2007             a           acer_campestre_subsp._campestre
#> 21  Baseline      2007             a                           acer_granatense
#> 22  Baseline      2007             a                          acer_heldreichii
#> 23  Baseline      2007             a                             acer_hyrcanum
#> 24  Baseline      2007             a                       acer_monspessulanum
#> 25  Baseline      2007             a acer_monspessulanum_subsp._monspessulanum
#> 26  Baseline      2007             a                              acer_negundo
#> 27  Baseline      2007             a                            acer_obtusatum
#> 28  Baseline      2007             a                               acer_opalus
#> 29  Baseline      2007             a                          acer_platanoides
#> 30  Baseline      2007             a                       acer_pseudoplatanus
#> 31  Baseline      2007             a                         acer_sempervirens
#> 32  Baseline      2007             a                            acer_tataricum
#> 33  Baseline      2007             a                         acer_trautvetteri
#> 34  Baseline      2007             a                          achillea_arabica
#> 35  Baseline      2007             a                    achillea_aspleniifolia
#> 36  Baseline      2007             a                           achillea_atrata
#> 37  Baseline      2007             a                        achillea_clavennae
#> 38  Baseline      2007             a                         achillea_clusiana
#> 39  Baseline      2007             a                       achillea_clypeolata
#> 40  Baseline      2007             a                        achillea_coarctata
#> 41  Baseline      2007             a                          achillea_collina
#> 42  Baseline      2007             a                     achillea_crithmifolia
#> 43  Baseline      2007             a                          achillea_distans
#> 44  Baseline      2007             a           achillea_distans_subsp._stricta
#> 45  Baseline      2007             a       achillea_erba-rotta_subsp._moschata
#> 46  Baseline      2007             a                      achillea_grandifolia
#> 47  Baseline      2007             a                      achillea_holosericea
#> 48  Baseline      2007             a                        achillea_ligustica
#> 49  Baseline      2007             a                      achillea_macrophylla
#> 50  Baseline      2007             a                         achillea_maritima
#> 51  Baseline      2007             a         achillea_maritima_subsp._maritima
#> 52  Baseline      2007             a                        achillea_micrantha
#> 53  Baseline      2007             a                    achillea_micranthoides
#> 54  Baseline      2007             a                      achillea_millefolium
#> 55  Baseline      2007             a                achillea_millefolium_aggr.
#> 56  Baseline      2007             a   achillea_millefolium_subsp._millefolium
#> 57  Baseline      2007             a      achillea_millefolium_subsp._sudetica
#> 58  Baseline      2007             a                             achillea_nana
#> 59  Baseline      2007             a                          achillea_nobilis
#> 60  Baseline      2007             a       achillea_nobilis_subsp._neilreichii
#> 61  Baseline      2007             a           achillea_nobilis_subsp._nobilis
#> 62  Baseline      2007             a                          achillea_odorata
#> 63  Baseline      2007             a                          achillea_oxyloba
#> 64  Baseline      2007             a           achillea_oxyloba_subsp._schurii
#> 65  Baseline      2007             a                        achillea_pannonica
#> 66  Baseline      2007             a                         achillea_ptarmica
#> 67  Baseline      2007             a                        achillea_pyrenaica
#> 68  Baseline      2007             a                      achillea_salicifolia
#> 69  Baseline      2007             a  achillea_santolinoides_subsp._wilhelmsii
#> 70  Baseline      2007             a                          achillea_setacea
#> 71  Baseline      2007             a                        achillea_tomentosa
#> 72  Baseline      2007             a                        achillea_virescens
#> 73  Baseline      2007             a                     achnatherum_bromoides
#> 74  Baseline      2007             a                 achnatherum_calamagrostis
#> 75  Baseline      2007             a                    aconitum_angustifolium
#> 76  Baseline      2007             a                          aconitum_anthora
#> 77  Baseline      2007             a                          aconitum_degenii
#> 78  Baseline      2007             a                           aconitum_firmum
#> 79  Baseline      2007             a                       aconitum_lycoctonum
#> 80  Baseline      2007             a     aconitum_lycoctonum_subsp._lycoctonum
#> 81  Baseline      2007             a   aconitum_lycoctonum_subsp._neapolitanum
#> 82  Baseline      2007             a       aconitum_lycoctonum_subsp._vulparia
#> 83  Baseline      2007             a                       aconitum_moldavicum
#> 84  Baseline      2007             a                         aconitum_napellus
#> 85  Baseline      2007             a                   aconitum_napellus_aggr.
#> 86  Baseline      2007             a          aconitum_napellus_subsp._vulgare
#> 87  Baseline      2007             a                         aconitum_plicatum
#> 88  Baseline      2007             a                         aconitum_tauricum
#> 89  Baseline      2007             a                       aconitum_variegatum
#> 90  Baseline      2007             a                        aconogonon_alpinum
#> 91  Baseline      2007             a                            acorus_calamus
#> 92  Baseline      2007             a                           actaea_europaea
#> 93  Baseline      2007             a                            actaea_spicata
#> 94  Baseline      2007             a                   adenocarpus_complicatus
#> 95  Baseline      2007             a                    adenocarpus_telonensis
#> 96  Baseline      2007             a                     adenostyles_alliariae
#> 97  Baseline      2007             a    adenostyles_alliariae_subsp._alliariae
#> 98  Baseline      2007             a                        adenostyles_alpina
#> 99  Baseline      2007             a          adenostyles_alpina_subsp._alpina
#> 100 Baseline      2007             a       adenostyles_alpina_subsp._pyrenaica
#> 101 Baseline      2007             a          adenostyles_glabra_subsp._glabra
#> 102 Baseline      2007             a                   adenostyles_leucophylla
#> 103 Baseline      2007             a                 adiantum_capillus-veneris
#> 104 Baseline      2007             a                         adonis_aestivalis
#> 105 Baseline      2007             a                              adonis_annua
#> 106 Baseline      2007             a                            adonis_flammea
#> 107 Baseline      2007             a                         adonis_microcarpa
#> 108 Baseline      2007             a                           adonis_vernalis
#> 109 Baseline      2007             a                       adoxa_moschatellina
#> 110 Baseline      2007             a                       aegilops_biuncialis
#> 111 Baseline      2007             a                          aegilops_caudata
#> 112 Baseline      2007             a                       aegilops_cylindrica
#> 113 Baseline      2007             a                       aegilops_geniculata
#> 114 Baseline      2007             a                         aegilops_neglecta
#> 115 Baseline      2007             a                      aegilops_triuncialis
#> 116 Baseline      2007             a   aegilops_triuncialis_subsp._triuncialis
#> 117 Baseline      2007             a                      aegilops_umbellulata
#> 118 Baseline      2007             a               aegonychon_purpurocaeruleum
#> 119 Baseline      2007             a                     aegopodium_podagraria
#> 120 Baseline      2007             a                      aeluropus_littoralis
#> 121 Baseline      2007             a                    aesculus_hippocastanum
#> 122 Baseline      2007             a                       aethionema_arabicum
#> 123 Baseline      2007             a                       aethionema_saxatile
#> 124 Baseline      2007             a        aethionema_saxatile_subsp._graecum
#> 125 Baseline      2007             a                          aethusa_cynapium
#> 126 Baseline      2007             a          aethusa_cynapium_subsp._cynapium
#> 127 Baseline      2007             a                       agrimonia_eupatoria
#> 128 Baseline      2007             a      agrimonia_eupatoria_subsp._eupatoria
#> 129 Baseline      2007             a                         agrimonia_procera
#> 130 Baseline      2007             a                       agropyron_cristatum
#> 131 Baseline      2007             a       agropyron_cristatum_subsp._brandzae
#> 132 Baseline      2007             a     agropyron_cristatum_subsp._pectinatum
#> 133 Baseline      2007             a                      agropyron_dasyanthum
#> 134 Baseline      2007             a                        agrostemma_githago
#> 135 Baseline      2007             a                           agrostis_alpina
#> 136 Baseline      2007             a                           agrostis_canina
#> 137 Baseline      2007             a                     agrostis_canina_aggr.
#> 138 Baseline      2007             a                      agrostis_canina_s.l.
#> 139 Baseline      2007             a                       agrostis_capillaris
#> 140 Baseline      2007             a                       agrostis_castellana
#> 141 Baseline      2007             a                         agrostis_curtisii
#> 142 Baseline      2007             a                         agrostis_gigantea
#> 143 Baseline      2007             a         agrostis_gigantea_subsp._maeotica
#> 144 Baseline      2007             a                        agrostis_hesperica
#> 145 Baseline      2007             a                        agrostis_mertensii
#> 146 Baseline      2007             a                        agrostis_rupestris
#> 147 Baseline      2007             a                      agrostis_schleicheri
#> 148 Baseline      2007             a                     agrostis_schraderiana
#> 149 Baseline      2007             a                      agrostis_stolonifera
#> 150 Baseline      2007             a                agrostis_stolonifera_aggr.
#> 151 Baseline      2007             a      agrostis_stolonifera_subsp._maritima
#> 152 Baseline      2007             a                         agrostis_vinealis
#> 153 Baseline      2007             a                   agrostis_x_fouilladeana
#> 154 Baseline      2007             a                       ailanthus_altissima
#> 155 Baseline      2007             a                         aira_caryophyllea
#> 156 Baseline      2007             a     aira_caryophyllea_subsp._caryophyllea
#> 157 Baseline      2007             a                            aira_cupaniana
#> 158 Baseline      2007             a                        aira_elegantissima
#> 159 Baseline      2007             a                              aira_praecox
#> 160 Baseline      2007             a                              aira_tenorei
#> 161 Baseline      2007             a                          aira_uniaristata
#> 162 Baseline      2007             a                         ajuga_chamaepitys
#> 163 Baseline      2007             a             ajuga_chamaepitys_subsp._chia
#> 164 Baseline      2007             a                          ajuga_genevensis
#> 165 Baseline      2007             a                                 ajuga_iva
#> 166 Baseline      2007             a                           ajuga_laxmannii
#> 167 Baseline      2007             a                          ajuga_orientalis
#> 168 Baseline      2007             a                         ajuga_pyramidalis
#> 169 Baseline      2007             a                             ajuga_reptans
#> 170 Baseline      2007             a                             alcea_biennis
#> 171 Baseline      2007             a                      alchemilla_acutiloba
#> 172 Baseline      2007             a                       alchemilla_alpigena
#> 173 Baseline      2007             a                         alchemilla_alpina
#> 174 Baseline      2007             a                       alchemilla_anisiaca
#> 175 Baseline      2007             a                    alchemilla_catalaunica
#> 176 Baseline      2007             a                      alchemilla_caucasica
#> 177 Baseline      2007             a                       alchemilla_colorata
#> 178 Baseline      2007             a                      alchemilla_conjuncta
#> 179 Baseline      2007             a                alchemilla_conjuncta_aggr.
#> 180 Baseline      2007             a                       alchemilla_coriacea
#> 181 Baseline      2007             a                 alchemilla_coriacea_aggr.
#> 182 Baseline      2007             a                        alchemilla_crinita
#> 183 Baseline      2007             a                     alchemilla_filicaulis
#> 184 Baseline      2007             a                          alchemilla_fissa
#> 185 Baseline      2007             a                     alchemilla_flabellata
#> 186 Baseline      2007             a                         alchemilla_glabra
#> 187 Baseline      2007             a                   alchemilla_glabra_aggr.
#> 188 Baseline      2007             a                    alchemilla_glaucescens
#> 189 Baseline      2007             a                       alchemilla_hoppeana
#> 190 Baseline      2007             a                        alchemilla_hybrida
#> 191 Baseline      2007             a                  alchemilla_hybrida_aggr.
#> 192 Baseline      2007             a                         alchemilla_micans
#> 193 Baseline      2007             a                      alchemilla_monticola
#> 194 Baseline      2007             a                   alchemilla_pentaphyllea
#> 195 Baseline      2007             a       alchemilla_plicatula_var._plicatula
#> 196 Baseline      2007             a                      alchemilla_saxatilis
#> 197 Baseline      2007             a                     alchemilla_subcrenata
#> 198 Baseline      2007             a                       alchemilla_vulgaris
#> 199 Baseline      2007             a                 alchemilla_vulgaris_aggr.
#> 200 Baseline      2007             a                   alchemilla_xanthochlora
#>     distance
#> 1   3.636952
#> 2   3.406591
#> 3   3.123997
#> 4   2.899396
#> 5   3.219113
#> 6   3.387160
#> 7   3.683923
#> 8   2.860623
#> 9   2.913475
#> 10  2.833773
#> 11  2.809365
#> 12  2.937199
#> 13  2.802648
#> 14  2.850136
#> 15  3.068189
#> 16  3.251038
#> 17  3.257391
#> 18  3.277638
#> 19  3.316213
#> 20  3.126469
#> 21  3.012411
#> 22  3.591992
#> 23  3.231464
#> 24  3.140927
#> 25  3.287096
#> 26  3.007178
#> 27  3.444014
#> 28  3.297285
#> 29  3.406773
#> 30  3.413035
#> 31  3.212860
#> 32  3.182642
#> 33  3.599463
#> 34  2.830811
#> 35  2.830588
#> 36  3.653152
#> 37  3.598935
#> 38  3.314449
#> 39  2.927989
#> 40  2.901220
#> 41  2.868363
#> 42  2.878634
#> 43  3.238368
#> 44  3.191846
#> 45  3.681318
#> 46  3.079541
#> 47  3.103055
#> 48  2.942638
#> 49  3.971991
#> 50  3.234618
#> 51  3.259122
#> 52  2.801288
#> 53  2.819302
#> 54  2.842956
#> 55  2.889183
#> 56  2.797223
#> 57  3.297393
#> 58  3.586640
#> 59  2.849483
#> 60  2.851591
#> 61  2.854779
#> 62  2.916762
#> 63  3.569481
#> 64  3.336302
#> 65  2.851478
#> 66  2.858622
#> 67  3.039153
#> 68  2.838002
#> 69  2.832791
#> 70  2.868584
#> 71  2.985516
#> 72  3.094862
#> 73  3.189521
#> 74  3.067461
#> 75  3.640520
#> 76  3.046668
#> 77  3.904280
#> 78  3.365457
#> 79  3.800489
#> 80  3.588196
#> 81  3.562313
#> 82  3.808958
#> 83  3.520008
#> 84  3.365612
#> 85  3.831748
#> 86  3.509561
#> 87  3.312191
#> 88  3.522273
#> 89  3.374926
#> 90  3.168574
#> 91  2.883851
#> 92  3.311772
#> 93  3.610888
#> 94  2.997037
#> 95  3.065979
#> 96  3.744186
#> 97  3.886905
#> 98  3.881803
#> 99  3.547925
#> 100 3.555039
#> 101 3.797002
#> 102 3.338009
#> 103 3.240927
#> 104 2.865460
#> 105 3.046768
#> 106 2.869221
#> 107 3.013274
#> 108 2.902688
#> 109 3.442337
#> 110 3.122875
#> 111 3.250102
#> 112 2.892887
#> 113 3.021595
#> 114 2.933335
#> 115 2.947133
#> 116 2.963894
#> 117 2.932909
#> 118 3.178334
#> 119 3.225339
#> 120 2.967920
#> 121 3.280604
#> 122 2.843364
#> 123 3.028566
#> 124 3.058361
#> 125 2.883293
#> 126 2.826450
#> 127 2.887922
#> 128 2.825901
#> 129 2.949042
#> 130 2.888303
#> 131 2.904341
#> 132 2.865567
#> 133 2.788287
#> 134 2.844972
#> 135 3.529333
#> 136 2.829460
#> 137 2.847522
#> 138 3.058093
#> 139 2.836853
#> 140 2.885383
#> 141 2.855063
#> 142 2.828203
#> 143 2.774186
#> 144 2.771401
#> 145 3.144778
#> 146 3.313447
#> 147 3.046642
#> 148 3.467704
#> 149 2.846342
#> 150 2.883735
#> 151 3.002495
#> 152 2.810356
#> 153 2.924155
#> 154 2.985830
#> 155 2.837323
#> 156 2.825524
#> 157 3.103830
#> 158 3.006436
#> 159 2.800826
#> 160 3.078650
#> 161 3.064657
#> 162 2.922216
#> 163 2.871643
#> 164 2.890729
#> 165 3.114191
#> 166 2.919480
#> 167 2.885406
#> 168 3.193739
#> 169 3.203512
#> 170 3.695590
#> 171 2.965728
#> 172 3.122777
#> 173 3.346503
#> 174 3.621825
#> 175 3.250545
#> 176 3.351831
#> 177 3.183223
#> 178 3.739237
#> 179 3.629804
#> 180 3.287342
#> 181 3.451742
#> 182 3.101606
#> 183 3.077561
#> 184 3.398253
#> 185 3.170051
#> 186 3.114937
#> 187 3.511190
#> 188 2.895560
#> 189 3.364724
#> 190 3.126706
#> 191 3.412525
#> 192 2.966234
#> 193 2.971383
#> 194 3.579574
#> 195 3.366188
#> 196 3.119380
#> 197 2.999995
#> 198 3.050119
#> 199 3.031662
#> 200 3.029799
#>  [ reached 'max' / getOption("max.print") -- omitted 6431 rows ]